[ghstack-poisoned]

+++ torch/csrc/inductor/aoti_runtime/thread_local.h
@@ -80,7 +80,7 @@ struct ThreadLocalCachedOutputTensor<ArrayRefTensor<T>> {
   }
 
   std::unique_ptr<T[]> storage_;
-  size_t capacity_ = 0;
+  int64_t capacity_ = 0;
   RAIIAtenTensorHandle tensor_;
 };
